#61 Moved post model to models folder.
